Moved exceptions to their own namespace.
