+++ src/Beta.TestAdapter/Exceptions/BetaControllerNotFoundException.cs
@@ -1,0 +1,8 @@
+﻿namespace Beta.TestAdapter.Exceptions;
+
+/// <summary>
+///     Defines an exception that is thrown when the Beta controller is not found.
+/// </summary>
+/// <param name="message">A message describing why the controller could not be found.</param>
+public class BetaControllerNotFoundException(string message)
+    : Exception(message);
