+++ src/Beta.TestAdapter/Exceptions/BetaEngineLoadFailedException.cs
@@ -1,0 +1,9 @@
+﻿namespace Beta.TestAdapter.Exceptions;
+
+/// <summary>
+///     Defines an exception that is thrown when the controller fails to load.
+/// </summary>
+/// <param name="message">A message explaining why the controller failed to load.</param>
+/// <param name="innerException">The inner exception.</param>
+public class BetaEngineLoadFailedException(string message, Exception innerException)
+    : Exception(message, innerException);
